update check permission logic
